Disable unit tests for master branch
Currently, master branch in cinderlib is used for development of yoga
release (it's trailing-cycle project). After last changes in cinder
project, cinderlib in master is not compatible with cinder master and
fails unit tests, so i'm disabling unit tests until yoga release is
created and master branch compatible with zed.

Change-Id: I09da6d6900f75d7103e7d6e1b4d52bae87753773